+++ apis/python/HISTORY.md
@@ -12,6 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Deprecated
 
+- \[[#4275](https://github.com/single-cell-data/TileDB-SOMA/pull/4275)\] Deprecate allowing a dimension in `shape` for a new `SOMASpaseNDArray` to be defaulted to `1` with `None` in the `create` method. In the future, the shape must be a sequence of positive integers.
+- \[[#4275](https://github.com/single-cell-data/TileDB-SOMA/pull/4275)\] Deprecate leaving `domain=None` when creating a `SOMADataFrame`, `SOMAPointCloudDataFrame`, or `SOMAGeometryDataFrame`. In the future, the domain must be fully specified on creation.
+
 ### Removed
 
 ### Fixed